Fix issue with high volume of data to clients
The websocket server was previously sending a high volume of data to clients when refreshing the frontend client multiple times. I initially had thought this was due to the server maintaining connections even after refreshing.

It turns out that the previous code created event listeners for the updateLeaderboard event everytime a new connection was created, not just a single time upon initialization of the server. This led to multiple listeners and emitters each time a connection was made.

I naively thought that everything had to be nested within the server connection event block.

Also, fix typing for error handling.